pyproject.toml
[tool.poetry]
name = "nodeman"
version = "0.0.0"
description = "DNS TAPIR Node Manager"
authors = ["Jakob Schlyter <jakob@kirei.se>"]
readme = "README.md"

[tool.poetry.scripts]
nodeman_server = "nodeman.server:main"
nodeman_client = "nodeman.client:main"

[tool.poetry.dependencies]
python = "^3.12"
dnstapir = {git = "https://github.com/dnstapir/python-dnstapir.git", rev = "v1.0.1", extras = ["opentelemetry"]}
mongoengine = "^0.29.0"
fastapi = ">=0.114.0"
uvicorn = ">=0.30.1"
pydantic-settings = "^2.6.0"
werkzeug = "^3.0.4"
pyyaml = "^6.0.1"
namesgenerator = "^0.3"
jwcrypto = "^1.5.6"
httpx = "^0.27.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
ruff = ">=0.8.0"
pytest-ruff = "^0.4.1"
pytest-asyncio = ">=0.24.0"
pytest-httpx = ">=0.32.0"
mongomock = "^4.3.0"
coverage = "^7.6.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = ["E501", "I001"]

[tool.ruff.lint.isort]
force-sort-within-sections = false
known-first-party = [
    "nodeman", "dnstapir"
]
extra-standard-library = [
    "tomllib",
]
combine-as-imports = true
split-on-trailing-comma = false
